Unclobber mutt_parse headers
