+++ mutt_parse.c
@@ -32,11 +32,14 @@
 #include <stdio.h>
 #include "mutt/mutt.h"
 #include "email/lib.h"
-#include "mutt.h"
 #include "mutt_parse.h"
-#include "globals.h"
 #include "mx.h"
 #include "ncrypt/ncrypt.h"
+
+struct ListHead AttachAllow = STAILQ_HEAD_INITIALIZER(AttachAllow);     ///< List of attachment types to be counted
+struct ListHead AttachExclude = STAILQ_HEAD_INITIALIZER(AttachExclude); ///< List of attachment types to be ignored
+struct ListHead InlineAllow = STAILQ_HEAD_INITIALIZER(InlineAllow);     ///< List of inline types to counted
+struct ListHead InlineExclude = STAILQ_HEAD_INITIALIZER(InlineExclude); ///< List of inline types to ignore
 
 /**
  * mutt_parse_mime_message - Parse a MIME email
